fix signed prange index in OpenMP

+++ pulse2percept/models/_horsager2009.pyx
@@ -63,7 +63,7 @@ cpdef temporal_fast(const float32[:, ::1] stim,
         float32 ca, r1, r2, r3, r4a, r4b, r4c
         float32 t_sim, amp
         float32[:, ::1] percept
-        uint32 idx_space, idx_sim, idx_stim, idx_frame
+        size_t idx_space, idx_sim, idx_stim, idx_frame
         uint32 n_space, n_stim, n_percept, n_sim
 
     # Note that eps must be divided by 1000, because the original model was fit
